separate interpeter stage
