+++ src/types/resolveStatements.ts
@@ -27,8 +27,8 @@ import {
 } from "./resolveDescriptors";
 import { getExpType, resolveExpression } from "./resolveExpression";
 import { FunctionDescription, printTypeRef, TypeRef } from "./types";
-import { evalConstantExpression } from "../constEval";
-import { ensureInt } from "../interpreter";
+import { evalConstantExpression } from "../interpreter/constEval";
+import { ensureInt } from "../interpreter/interpreter";
 import { crc16 } from "../utils/crc16";
 import { SrcInfo } from "../grammar";
